Move uneeded files to dump
